# docs/swagger.yaml
definitions:
  handler.ConfirmLoginResp:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.CreateEventID:
    properties:
      event_id:
        type: integer
    type: object
  handler.CreateEventResp:
    properties:
      data:
        $ref: '#/definitions/handler.CreateEventID'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.Event:
    properties:
      current_player:
        items:
          $ref: '#/definitions/handler.UserInfo'
        type: array
      current_player_num:
        type: integer
      end_time:
        type: integer
      end_time_str:
        type: string
      field_name:
        type: string
      game_type:
        type: string
      highest_level:
        type: number
      is_booked:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      lowest_level:
        type: number
      match_type:
        type: string
      max_player_num:
        type: integer
      start_time:
        type: integer
      start_time_str:
        type: string
      status:
        description: event状态：created;full;finished
        type: string
      weekday:
        type: string
    type: object
  handler.EventTab:
    properties:
      event_info:
        $ref: '#/definitions/model.EventInfo'
      players:
        items:
          $ref: '#/definitions/model.Player'
        type: array
    type: object
  handler.GetApprovalItemsResp:
    properties:
      data:
        $ref: '#/definitions/handler.GetApprovalItemsResult'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetApprovalItemsResult:
    properties:
      calibration_items:
        description: 待审批的定位为职业的用户列表
        items:
          $ref: '#/definitions/mysql.WechatUserInfoWithoutGorm'
        type: array
      organization_items:
        description: 待审批的组织列表
        items:
          $ref: '#/definitions/mysql.OrganizationWithoutGorm'
        type: array
    type: object
  handler.GetCalibrationQuestionsResp:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetEventListData:
    properties:
      event_list:
        items:
          $ref: '#/definitions/handler.Event'
        type: array
      has_more:
        type: boolean
      offset:
        type: integer
    type: object
  handler.GetEventListResp:
    properties:
      data:
        $ref: '#/definitions/handler.GetEventListData'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetFieldsListRes:
    properties:
      fields_list:
        items:
          $ref: '#/definitions/handler.Organization'
        type: array
    type: object
  handler.GetFieldsListResp:
    properties:
      data:
        $ref: '#/definitions/handler.GetFieldsListRes'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetHostInfoResp:
    properties:
      data:
        additionalProperties:
          $ref: '#/definitions/handler.Info'
        type: object
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetMyTabResp:
    properties:
      data:
        $ref: '#/definitions/handler.MyTab'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetScoreResultResp:
    properties:
      data:
        $ref: '#/definitions/handler.MatchResult'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.GetScoreboardResp:
    properties:
      data:
        $ref: '#/definitions/handler.ScoreBoard'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.Info:
    properties:
      is_host:
        type: boolean
      organization_id:
        type: integer
    type: object
  handler.LevelChange:
    properties:
      change:
        type: number
      date:
        type: string
    type: object
  handler.LevelInfo:
    properties:
      current_level:
        type: number
      level_detail:
        items:
          $ref: '#/definitions/handler.LevelChange'
        type: array
      status:
        description: 定级状态：need_to_calibrate;wait_for_approve;approved(7.0以下自动审批)
        type: string
    type: object
  handler.MatchDetail:
    properties:
      round_info:
        items:
          $ref: '#/definitions/handler.RoundInfo'
        type: array
      round_target_score:
        description: 目标得分
        type: integer
      settings:
        $ref: '#/definitions/handler.Settings'
    type: object
  handler.MatchResult:
    properties:
      players_detail:
        items:
          $ref: '#/definitions/handler.PlayerAfterMatch'
        type: array
      round_detail:
        items:
          $ref: '#/definitions/handler.UploadRoundInfo'
        type: array
    type: object
  handler.MyTab:
    properties:
      sport_type_infos:
        items:
          $ref: '#/definitions/handler.SportTypeInfo'
        type: array
      user_info:
        $ref: '#/definitions/handler.UserInfo'
    type: object
  handler.Organization:
    properties:
      address:
        type: string
      event_num:
        type: integer
      id:
        type: integer
      latitude:
        type: number
      logo:
        type: string
      longitude:
        type: number
      name:
        type: string
      status:
        description: 组织状态：no_organization;wait_for_approve;approved
        type: string
    type: object
  handler.PlayerAfterMatch:
    properties:
      avatar:
        type: string
      is_calibrated:
        type: boolean
      level:
        type: number
      level_change:
        description: 等级的变化,不对外返回
        type: number
      level_change_str:
        description: 对外返回的等级变化
        type: string
      lose_round:
        type: integer
      nick_name:
        type: string
      open_id:
        type: string
      rank:
        description: 排名
        type: integer
      tie_round:
        type: integer
      total_score:
        description: 总得分
        type: integer
      win_round:
        type: integer
    type: object
  handler.RoundInfo:
    properties:
      away:
        description: 客队球员
        items:
          $ref: '#/definitions/model.Player'
        type: array
      away_avg:
        description: 客队平均分
        type: number
      court_num:
        description: 场地号
        type: integer
      home:
        description: 主队球员
        items:
          $ref: '#/definitions/model.Player'
        type: array
      home_avg:
        description: 主队平均分
        type: number
      round_num:
        description: 轮次数
        type: integer
    type: object
  handler.ScoreBoard:
    properties:
      event_info:
        $ref: '#/definitions/model.EventInfo'
      options:
        $ref: '#/definitions/model.ScoreOptions'
      players:
        items:
          $ref: '#/definitions/model.Player'
        type: array
    type: object
  handler.Settings:
    properties:
      valid_scorers:
        items:
          $ref: '#/definitions/model.Player'
        type: array
    type: object
  handler.SportTypeInfo:
    properties:
      level_info:
        allOf:
        - $ref: '#/definitions/handler.LevelInfo'
        description: 级别信息
      my_games:
        items:
          $ref: '#/definitions/handler.Event'
        type: array
      my_organization:
        $ref: '#/definitions/handler.Organization'
      sport_type:
        type: string
    type: object
  handler.StartScoringResp:
    properties:
      data:
        $ref: '#/definitions/handler.MatchDetail'
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  handler.UpdateUserInfoReq:
    properties:
      avatar_url:
        type: string
      encryptedData:
        description: 包括敏感数据在内的完整用户信息的加密数据
        type: string
      iv:
        description: 加密算法的初始向量
        type: string
      nick_name:
        type: string
      rawData:
        description: 不包括敏感信息的原始数据字符串，用于计算签名
        type: string
      signature:
        description: 使用 sha1( rawData + sessionkey ) 得到字符串，用于校验用户信息
        type: string
    type: object
  handler.UploadRoundInfo:
    properties:
      away:
        description: 客队球员
        items:
          $ref: '#/definitions/model.Player'
        type: array
      away_avg:
        description: 客队平均分
        type: number
      away_level_change:
        type: number
      away_score:
        description: 客队本局分数
        type: integer
      court_num:
        description: 场地号
        type: integer
      home:
        description: 主队球员
        items:
          $ref: '#/definitions/model.Player'
        type: array
      home_avg:
        description: 主队平均分
        type: number
      home_level_change:
        type: number
      home_score:
        description: 主队本局分数
        type: integer
      round_num:
        description: 轮次数
        type: integer
      winner:
        description: 获胜者
        type: string
    type: object
  handler.UserInfo:
    properties:
      avatar_url:
        type: string
      nick_name:
        type: string
    type: object
  handler.UserLoginResp:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  model.BackEndResp:
    properties:
      data: {}
      err_no:
        type: integer
      err_tips:
        type: string
    type: object
  model.EventInfo:
    description: 活动元信息
    properties:
      city:
        type: string
      creator_avatar:
        description: 创建者头像
        type: string
      creator_nickname:
        description: 创建者昵称
        type: string
      current_people:
        type: integer
      date:
        description: 日期 20060102
        type: string
      desc:
        type: string
      end_time:
        type: integer
      end_time_str:
        type: string
      field_name:
        type: string
      field_type:
        description: 场地类型
        type: string
      game_type:
        description: 对局类型，solo/duo
        type: string
      highest_level:
        description: 适合的最高级别
        type: number
      id:
        description: 数据库主键
        type: integer
      is_booked:
        description: 是否已定场
        type: boolean
      is_competitive:
        description: 是否是竞赛类型
        type: boolean
      is_draft:
        description: 是否是草稿请求
        type: boolean
      is_host:
        description: 是否是组织创建
        type: boolean
      is_public:
        description: 是否是公开比赛
        type: boolean
      latitude:
        type: string
      longitude:
        type: string
      lowest_level:
        description: 适合的最低级别
        type: number
      max_people_num:
        type: integer
      name:
        type: string
      organization_id:
        description: 组织ID
        type: integer
      organization_logo:
        type: string
      price:
        type: integer
      self_join:
        description: 自己是否加入
        type: boolean
      sport_type:
        description: 运动类型
        type: string
      start_time:
        type: integer
      start_time_str:
        type: string
      status:
        description: full, can_join, finished
        type: string
      weekday:
        description: 星期几
        type: string
    type: object
  model.Player:
    properties:
      avatar:
        type: string
      is_calibrated:
        type: boolean
      level:
        type: number
      nick_name:
        type: string
      open_id:
        type: string
    type: object
  model.Question:
    properties:
      options:
        additionalProperties:
          type: string
        type: object
      q_id:
        type: integer
      q_text:
        type: string
    type: object
  model.ScoreOptions:
    properties:
      available_round_target:
        additionalProperties:
          items:
            type: integer
          type: array
        description: 可选的每轮目标分
        type: object
      available_score_rule:
        description: 可选的记分赛制
        items:
          type: string
        type: array
      field_num:
        description: 比赛场地数
        type: integer
    type: object
  mysql.OrganizationWithoutGorm:
    properties:
      address:
        description: 详细地址
        type: string
      city:
        description: 城市
        type: string
      contact:
        description: 联系方式
        type: string
      host_open_id:
        description: 主理人openID
        type: string
      is_approved:
        description: 是否通过审批
        type: integer
      logo:
        description: 组织图标logo
        type: string
      name:
        description: 组织名字
        type: string
      reviewer:
        description: 审批人
        type: string
      sport_type:
        description: 运动类型
        type: string
      total_event_num:
        description: 活动次数
        type: integer
    type: object
  mysql.WechatUserInfoWithoutGorm:
    properties:
      avatar:
        description: 微信用户头像
        type: string
      calibration_proof:
        description: 自称是pro的（定级7.0的人）需要额外提供图片
        type: string
      gender:
        description: 性别
        type: integer
      is_calibrated:
        description: 是否完成定级
        type: integer
      is_host:
        description: 1标识有，0标识无
        type: integer
      is_primary:
        description: 是否是主要的recprd
        type: integer
      joined_event:
        description: 参与的活动
        type: string
      joined_group:
        description: 参加的组织
        type: string
      joined_times:
        description: 参与次数
        type: integer
      level:
        description: 级别
        type: integer
      nickname:
        description: 微信用户昵称
        type: string
      open_id:
        description: 微信用户open_id
        type: string
      organization_id:
        description: 主理的组织名
        type: integer
      phone_number:
        description: 用户手机号
        type: integer
      preference:
        type: string
      reviewer:
        description: 审批人（只有职业才需要）
        type: string
      session_key:
        description: 微信session_key 用于解密
        type: string
      sport_type:
        description: 运动类型
        type: string
      tags:
        type: string
      union_id:
        description: 微信用户union_id
        type: string
    type: object
info:
  contact: {}
paths:
  /team_up/admin/approve:
    post:
      consumes:
      - application/json
      description: 管理员审批：包含创建组织的申请与Pro级别的认证事件
      parameters:
      - description: 审批事件类型：organization 或者 calibration_proof
        in: query
        name: approve_type
        required: true
        type: string
      - description: 要批准的组织ID，当审批事件类型为organization时必传
        in: body
        name: organization_id
        schema:
          type: integer
      - description: 要批准的的用户open_id，当审批事件类型为calibration_proof时必传
        in: body
        name: open_id
        schema:
          type: string
      - description: 要批准的用户运动类型，当审批事件类型为calibration_proof时必传
        in: body
        name: sport_type
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 管理员审批
      tags:
      - /team_up/admin
  /team_up/admin/get_approval_items:
    get:
      consumes:
      - application/json
      description: 管理员审批：获取待审批的事项（包括定级别为pro级别的用户或者用户创建的组织信息）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetApprovalItemsResp'
      summary: 待审批事项清单
      tags:
      - /team_up/admin
  /team_up/event/create:
    post:
      consumes:
      - application/json
      description: 个人或者组织创建活动
      parameters:
      - description: 参考EventInfo Model
        in: body
        name: create_event
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateEventResp'
      summary: 创建活动
      tags:
      - /team_up/event
  /team_up/event/fields_list:
    post:
      consumes:
      - application/json
      description: 根据运动类型和城市获取场地列表
      parameters:
      - description: 城市
        in: body
        name: city
        required: true
        schema:
          type: string
      - description: 运动类型
        in: body
        name: sport_type
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetFieldsListResp'
      summary: 获取场地列表
      tags:
      - /team_up/event
  /team_up/event/list:
    post:
      consumes:
      - application/json
      description: 根据筛选条件获取活动列表
      parameters:
      - description: 运动类型：pedal, tennis, pickleball
        in: body
        name: sport_type
        required: true
        schema:
          type: string
      - description: 城市
        in: body
        name: city
        required: true
        schema:
          type: string
      - description: 开始时间，秒级时间戳
        in: body
        name: start_time
        required: true
        schema:
          type: integer
      - description: 获取的数量
        in: body
        name: num
        required: true
        schema:
          type: integer
      - description: 偏移量
        in: body
        name: offset
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetEventListResp'
      summary: 获取活动列表
      tags:
      - /team_up/event
  /team_up/event/page:
    post:
      consumes:
      - application/json
      description: 活动详情页，包含活动元信息和参与的用户信息
      parameters:
      - description: 活动ID
        in: body
        name: event_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.EventTab'
      summary: 获取活动详情页
      tags:
      - /team_up/event
  /team_up/event/update:
    post:
      consumes:
      - application/json
      description: 个人或者组织更新活动元信息
      parameters:
      - description: 更新活动入参,参考EventInfo
        in: body
        name: update_event
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateEventResp'
      summary: 更新活动元信息
      tags:
      - /team_up/event
  /team_up/organization/create:
    post:
      consumes:
      - application/json
      description: 用户上传组织信息
      parameters:
      - description: 组织名称
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: 运动类型
        in: body
        name: sport_type
        required: true
        schema:
          type: string
      - description: 城市
        in: body
        name: city
        required: true
        schema:
          type: string
      - description: 地址
        in: body
        name: address
        required: true
        schema:
          type: string
      - description: 联系方式
        in: body
        name: contact
        required: true
        schema:
          type: string
      - description: logo图
        in: body
        name: logo
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 创建组织
      tags:
      - /team_up/organization
  /team_up/user/calibrate:
    post:
      consumes:
      - application/json
      description: 获取定级问题详情
      parameters:
      - description: 运动类型
        in: formData
        name: sport_type
        required: true
        type: string
      - description: 问卷结构体
        in: formData
        name: questionnaire
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetCalibrationQuestionsResp'
      summary: 用户定级
      tags:
      - /team_up/user
  /team_up/user/confirm_login:
    post:
      consumes:
      - application/json
      description: 前端使用微信code+获取手机号的code请求服务端登录
      parameters:
      - description: 静默登录的code
        in: body
        name: silent_code
        required: true
        schema:
          type: string
      - description: 获取电话号的code
        in: body
        name: phone_code
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ConfirmLoginResp'
      summary: 用户登录+获取手机号
      tags:
      - /team_up/user
  /team_up/user/get_calibration_questions:
    post:
      consumes:
      - application/json
      description: 获取定级问题详情
      parameters:
      - description: 获取定级问题入参
        in: body
        name: sport_type
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetCalibrationQuestionsResp'
      summary: 获取定级问题
      tags:
      - /team_up/user
  /team_up/user/get_host_info:
    get:
      consumes:
      - application/json
      description: 用户在不同运动类型下是否为“组织”身份
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetHostInfoResp'
      summary: 用户组织信息
      tags:
      - /team_up/user
  /team_up/user/get_score_result:
    post:
      consumes:
      - application/json
      description: 用户上传分数信息，服务端计算用户等级变化
      parameters:
      - description: 参考UploadRoundInfos
        in: body
        name: get_score_result
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetScoreResultResp'
      summary: 获取记分结果
      tags:
      - /team_up/user
  /team_up/user/get_scoreboard:
    post:
      consumes:
      - application/json
      description: 记分板页面，包含用户可选规则项
      parameters:
      - description: 活动ID
        in: body
        name: event_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetScoreboardResp'
      summary: 记分板页面
      tags:
      - /team_up/user
  /team_up/user/join_event:
    post:
      consumes:
      - application/json
      description: 加入活动场次
      parameters:
      - description: 加入活动场次入参
        in: body
        name: event_id
        required: true
        schema:
          type: integer
      - description: 是否通过邀请链接加入
        in: body
        name: is_inviting
        required: true
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 加入活动场次
      tags:
      - /team_up/user
  /team_up/user/login:
    post:
      consumes:
      - application/json
      description: 前端使用微信code请求服务端登录
      parameters:
      - description: 微信Code
        in: body
        name: code
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserLoginResp'
      summary: 用户登录(废弃，请使用confirm_login)
      tags:
      - /team_up/user
  /team_up/user/my_tab:
    get:
      consumes:
      - application/json
      description: 包含用户信息，级别信息和参与的活动
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetMyTabResp'
      summary: 我的页面
      tags:
      - /team_up/user
  /team_up/user/publish_score:
    post:
      consumes:
      - application/json
      description: 根据计算出的等级变化，服务端更新场次，用户的信息
      parameters:
      - description: 活动ID
        in: body
        name: event_id
        required: true
        schema:
          type: integer
      - description: 用户详情, 参考PublishScoreBody
        in: body
        name: player_detail
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 发布比赛结果
      tags:
      - /team_up/user
  /team_up/user/quit_event:
    post:
      consumes:
      - application/json
      description: 退出活动场次
      parameters:
      - description: 活动ID
        in: body
        name: event_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 退出活动场次
      tags:
      - /team_up/user
  /team_up/user/start_scoring:
    post:
      consumes:
      - application/json
      description: 根据用户选择的规则下发对应的详细对局信息
      parameters:
      - description: 用户选择的记分规则，参考StartScoringBody
        in: body
        name: start_scoring
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StartScoringResp'
      summary: 记分详细规则
      tags:
      - /team_up/user
  /team_up/user/update_phone_number:
    post:
      consumes:
      - application/json
      description: 前端获取加密的用户手机号，服务端进行解码，存储
      parameters:
      - description: 微信Code
        in: body
        name: code
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 获取用户手机号(废弃，请使用confirm_login)
      tags:
      - /team_up/user
  /team_up/user/update_user_info:
    post:
      consumes:
      - application/json
      description: 前端获取加密的用户手机号，服务端进行解码，存储
      parameters:
      - description: 请求
        in: body
        name: code
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 前端调用getUserProfile后调用，更新后端用户信息表
      tags:
      - /team_up/user
  /team_up/user/upload_image:
    post:
      consumes:
      - application/json
      description: 前端获取加密的用户手机号，服务端进行解码，存储
      parameters:
      - description: 图片类型:organization_logo, calibration_proof, event_image, avatar_image
        in: formData
        name: image_type
        required: true
        type: string
      - description: 文件流
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BackEndResp'
      summary: 前端上传文件流给服务端
      tags:
      - /team_up/user
swagger: "2.0"